requirements.txt.txt
telegram
python-telegram-bot==20.0
openai
requests
python-dotenv
yt-dlp